HBASE-11901 Improve the value size of the reference cell in mob column.(Jingcheng Du)

+++ hbase-server/src/main/java/org/apache/hadoop/hbase/mob/MobUtils.java
@@ -251,7 +251,7 @@ public class MobUtils {
     // find the original mob files by this table name. For details please see cloning
     // snapshot for mob files.
     existingTags.add(tableNameTag);
-    long valueLength = kv.getValueLength();
+    int valueLength = kv.getValueLength();
     byte[] refValue = Bytes.add(Bytes.toBytes(valueLength), fileName);
     KeyValue reference = new KeyValue(kv.getRowArray(), kv.getRowOffset(), kv.getRowLength(),
         kv.getFamilyArray(), kv.getFamilyOffset(), kv.getFamilyLength(), kv.getQualifierArray(),
